made some improvements to the messenger and removed the chat input from Messenger.svelte -> Messages.svelte(route)
