+++ frontend/src/routes/Messages.svelte
@@ -1,19 +1,21 @@
 <style lang="postcss">
+.max-h-messenger {
+  max-height: 45rem;
+}
 </style>
 
 <script lang="ts">
 //components import
 import DataTable from "../lib/table-elements/DataTable.svelte";
 import Text from "../lib/table-elements/Text.svelte";
-import Image from "../lib/table-elements/Image.svelte";
 import Messenger from "../lib/Messenger.svelte";
 import Avatar from "../lib/Avatar.svelte";
 //api
 import { getContacts } from "../api/contactsAxios";
+import { postNewMsg2Conversation } from "../api/messagesLogic";
 //stores
 import { availableContacts } from "../stores/contacts";
 import { selectedUser } from "../stores/messages";
-import { sha256 } from "../utils/sha256";
 //ecma imports
 import { onMount } from "svelte";
 import { io } from "socket.io-client";
@@ -21,7 +23,10 @@ import { io } from "socket.io-client";
 const backendUrl = import.meta.env.VITE_API_URL
   ? import.meta.env.VITE_API_URL
   : `https://api.${window.location.hostname}`;
-
+const userDomain =
+  import.meta.env.VITE_ENV_TYPE === "dev"
+    ? `did:web:localhost:3100`
+    : `did:web:api.${window.location.host}`;
 //TODO: move all of the socket logic in a separate ts or store file.
 
 let socket;
@@ -49,6 +54,36 @@ function openConversation(id: string) {
   selectedUser.set(id);
   socket.emit("joinRoom", $selectedUser);
 }
+
+//msg input DOM element
+let chatMsg: string;
+const onKeyPress = (e) => {
+  if (e.charCode === 13) {
+    newMsg();
+  }
+};
+
+async function newMsg() {
+  if (chatMsg) {
+    //building the new entry for the mongo db
+    const fullPayload: object = {
+      to: $selectedUser,
+      data: chatMsg,
+      when: new Date(),
+    };
+
+    //saving the data to the mongo db
+    await postNewMsg2Conversation(fullPayload);
+
+    //websocket new entry
+    socket.emit("chatToServer", {
+      sender: userDomain,
+      room: $selectedUser,
+      message: chatMsg,
+    });
+    chatMsg = "";
+  }
+}
 </script>
 
 <template>
@@ -56,12 +91,25 @@ function openConversation(id: string) {
     <!-- Off-canvas menu for mobile, show/hide based on off-canvas menu state. -->
     <div class="flex flex-col min-w-0 flex-1 overflow-hidden">
       <div class="flex-1 relative z-0 flex overflow-hidden">
-        <main
-          class="flex-1 relative z-0 overflow-y-auto focus:outline-none xl:order-last">
+        <main class="flex-1 relative z-0 focus:outline-none xl:order-last">
           <!-- Start main area-->
           <div class="absolute inset-0 py-6 px-4 sm:px-6 lg:px-8">
             {#if initMessenger}
-              <Messenger socket="{socket}" />
+              <!-- INFO: Where all the messages will displays-->
+              <div class="max-h-messenger overflow-y-auto">
+                <Messenger socket="{socket}" />
+              </div>
+              <!-- INFO: the chat input box where new messages are created-->
+              <div class="flex-grow min-w-full sticky">
+                <input
+                  bind:value="{chatMsg}"
+                  on:keypress="{onKeyPress}"
+                  type="text"
+                  name="msg-chat"
+                  id="msg-chat"
+                  class="flex-grow  fill-current shadow-sm focus:ring-indigo-500 focus:border-indigo-500 w-full sm:text-sm border-gray-300 rounded-md min-w-max"
+                  placeholder="Say hello!" />
+              </div>
             {/if}
           </div>
           <!-- End main area -->
@@ -98,4 +146,3 @@ function openConversation(id: string) {
     </div>
   </div>
 </template>
-<!-- clickable: "http://localhost:3000/" -->
