+++ frontend/tailwind.config.cjs
@@ -15,6 +15,13 @@ const config = {
         "1/2": "50%",
         "3/4": "75%",
       },
+      maxHeight: {
+        '0': '0',
+        '1/4': '25%',
+        '1/2': '50%',
+        '3/4': '75%',
+        'full': '100%',
+      }
     },
   },
   plugins: [require("@tailwindcss/forms")],
